+++ examples/full/main.go
@@ -2,8 +2,9 @@ package main
 
 import (
 	"fmt"
-	"github.com/AccentDesign/gcss/examples/full/styles"
 	"net/http"
+
+	"github.com/tomlkeate/gcss/examples/full/styles"
 )
 
 var (