+++ examples/full/styles/buttons.go
@@ -1,9 +1,9 @@
 package styles
 
 import (
-	"github.com/AccentDesign/gcss"
-	"github.com/AccentDesign/gcss/props"
-	"github.com/AccentDesign/gcss/variables"
+	"github.com/tomlkeate/gcss"
+	"github.com/tomlkeate/gcss/props"
+	"github.com/tomlkeate/gcss/variables"
 )
 
 // Buttons returns the styles for buttons for the base stylesheet.