+++ stylesheet.go
@@ -1,13 +1,17 @@
-package main
+package gcss
 
 import (
 	"context"
-	"github.com/tomlkeate/gcss"
+	"fmt"
 	"io"
 )
 
+type CSSComponent interface {
+	CSS(w io.Writer) error
+}
+
 // Stylesheet is a collection of styles.
-type Stylesheet []gcss.Style
+type Stylesheet []CSSComponent
 
 // Render writes the CSS representation of the stylesheet to the writer.
 // this is to ensure that it implements a templ.Component.
@@ -22,6 +26,26 @@ func (ss Stylesheet) Render(ctx context.Context, w io.Writer) error {
 		}
 	}
 	if _, err := io.WriteString(w, "\n</style>"); err != nil {
+		return err
+	}
+	return nil
+}
+
+type MediaBlock struct {
+	MediaQuery string
+	Styles     Stylesheet
+}
+
+func (mb *MediaBlock) CSS(w io.Writer) error {
+	if _, err := fmt.Fprintf(w, "%s{", mb.MediaQuery); err != nil {
+		return err
+	}
+	for _, style := range mb.Styles {
+		if err := style.CSS(w); err != nil {
+			return err
+		}
+	}
+	if _, err := fmt.Fprint(w, "}"); err != nil {
 		return err
 	}
 	return nil