+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+### v0.1.3
+
+Added stylesheets
+
+Added support for media queries
+
+Expanded props:
+
+- color hex support
+- color var support
+
 ### v0.1.2
 
 - Move the `Props` CSS writing into a `Props.CSS` func, so it can be used in isolation of the `Style` struct. This allows for the `Props` CSS to be used in other contexts, such as in a `style` attribute in HTML.